Rust #13561 When invoking "condor_status -server" the Disk totals were so large that it would wrap I updated the various total values to be "long" or "long long" as necessary I have tested and this solved the problem

+++ src/condor_status.V6/totals.h
@@ -103,8 +103,8 @@ class StartdRunTotal : public ClassTotal
 
 	protected:
 		int machines;
-		int condor_mips;
-		int kflops;
+		long condor_mips;
+		long kflops;
 		float loadavg;
 };
 
@@ -120,10 +120,10 @@ class StartdServerTotal : public ClassTotal
   	protected:
 		int machines;
 		int avail;
-		int memory;
-		unsigned long disk;
-		int condor_mips;
-		int kflops;
+		long memory;
+		unsigned long long disk;
+		long condor_mips;
+		long kflops;
 };
 
 
@@ -224,7 +224,7 @@ class CkptSrvrNormalTotal : public ClassTotal
 
 	protected:
 		int numServers;
-		unsigned long disk;
+		unsigned long long disk;
 };
 
 int getCODInt( ClassAd* ad, const char* id, const char* attr,
